Move functions and classes around
